ref: move inspector out of scraps
