+++ static/app/components/inspector.spec.tsx
