+++ static/app/components/inspector.tsx
